pulled and fixed eligibility tests from patient model test into eligibility testing file
